Code/Demo_Agentcore_Bedrock/requirements.txt
strands-agents
bedrock-agentcore
requests
beautifulsoup4
lxml
tenacity
playwright
